Rename auth_service.py to auth.py
